File upload

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const Product = require("../models/product");
 
+const fileHandler = require("../util/fileHandler");
+
 const { validationResult } = require("express-validator");
 
 exports.getAddProductsPage = (req, res, next) => {
@@ -17,7 +19,6 @@ exports.getAddProductsPage = (req, res, next) => {
     errorMessage: message,
     oldInput: {
       title: "",
-      imageURL: "",
       price: "",
       description: "",
     },
@@ -26,10 +27,26 @@ exports.getAddProductsPage = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageURL = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const valErrors = validationResult(req);
+
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      path: "/admin/add-product",
+      pageTitle: "Add product",
+      editing: false,
+      hasError: true,
+      errorMessage: "Invalid image format",
+      oldInput: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      validationErrors: [],
+    });
+  }
 
   if (!valErrors.isEmpty()) {
     console.log(valErrors.array());
@@ -41,17 +58,18 @@ exports.postAddProduct = (req, res, next) => {
       errorMessage: valErrors.array()[0].msg,
       oldInput: {
         title: title,
-        imageURL: imageURL,
+        imageUrl: imageUrl,
         price: price,
         description: description,
       },
       validationErrors: valErrors.array(),
     });
   }
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title,
-    imageUrl: imageURL,
+    imageUrl: imageUrl,
     price: price,
     description: description,
     userId: req.session.user,
@@ -78,6 +96,7 @@ exports.postAddProduct = (req, res, next) => {
       //   },
       //   validationErrors: [],
       // });
+      console.log("Error while creating a product");
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
@@ -111,7 +130,6 @@ exports.getEditProduct = (req, res, next) => {
         errorMessage: message,
         oldInput: {
           title: "",
-          imageURL: "",
           price: "",
           description: "",
         },
@@ -127,7 +145,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
-  const updatedImageURL = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
   const updatedPrice = req.body.price;
   const valErrors = validationResult(req);
@@ -137,25 +155,28 @@ exports.postEditProduct = (req, res, next) => {
     return res.status(422).render("admin/edit-product", {
       path: "/admin/add-product",
       pageTitle: "Edit product",
-      editing: false,
+      editing: true,
       hasError: true,
       errorMessage: valErrors.array()[0].msg,
       oldInput: {
         title: updatedTitle,
-        imageURL: updatedImageURL,
         price: updatedPrice,
         description: updatedDesc,
       },
       validationErrors: valErrors.array(),
     });
   }
+
   Product.findById(prodId)
     .then(product => {
       if (product.userId.toString() !== req.user._id.toString()) {
         return res.redirect("/");
       }
       product.title = updatedTitle;
-      product.imageUrl = updatedImageURL;
+      if (image) {
+        fileHandler.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       product.description = updatedDesc;
       product.price = updatedPrice;
       return product.save().then(() => {
@@ -190,7 +211,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.findByIdAndDelete(prodId)
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error("Product not found"));
+      }
+      fileHandler.deleteFile(product.imageUrl);
+      return Product.findByIdAndDelete(prodId);
+    })
     .then(product => {
       console.log(`Product Destroyed: \n ${product}`);
       res.redirect("/admin/products");
